refactor: split file/folder parsing from parseArguments
