+++ src/parseArguments.ts
@@ -1,13 +1,11 @@
 import version from "~/options/version";
 import help from "~/options/help";
-import parseFile from "~/parseFile";
-import { rmdirSync, mkdirSync, readdirSync } from "node:fs";
+import parseTarget from "~/parseTarget.ts";
 
 const OUTPUT_DIR_DEFAULT = "til";
 
 export default async function parseArguments(args: string[]) {
-  let foundTarget = false;
-  let target = "";
+  let target: string | null = null;
   let exit = false;
   let outputDir = OUTPUT_DIR_DEFAULT;
 
@@ -39,46 +37,12 @@ export default async function parseArguments(args: string[]) {
         }
         break;
       default:
-        foundTarget = true;
         target = arg;
     }
   }
 
-  if (foundTarget) {
-    const filesToProcess = [];
-    const isFolder = !(target.endsWith(".txt") || target.endsWith(".md"));
-
-    console.log(`Reading files...`);
-
-    if (isFolder) {
-      // Parse folder of text files
-      const dir = target;
-      let files = readdirSync(dir).filter(
-        (file) => file.endsWith(".txt") || file.endsWith(".md")
-      );
-      files = files.map((file) => `${dir}/${file}`);
-
-      files.forEach((file) => filesToProcess.push(file));
-    } else {
-      filesToProcess.push(target);
-    }
-
-    // Clear existing output and create folder
-    rmdirSync(`./${outputDir}`, { recursive: true });
-    mkdirSync(`./${outputDir}`, { recursive: true });
-
-    // Copy stylesheet
-    const stylesheet = Bun.file("src/style.css");
-    await Bun.write(`./${outputDir}/style.css`, stylesheet);
-
-    // Parse filesToProcess of files
-    for (const file of filesToProcess) {
-      await parseFile(file, outputDir);
-    }
-
-    console.log(
-      `Successfully generated ${filesToProcess.length} HTML file(s) to directory '${outputDir}'.`
-    );
+  if (!!target) {
+    parseTarget(target, outputDir);
   } else {
     throw new Error("Missing target files.");
   }
